+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 **New**
 - First pass for themes. [#38](https://github.com/69MichelleDB/ToNCodes/issues/38)
+- Added option to enable debug in File > Options.
 
 **Fixes**
 - Checking for updates when you were on the latest version, wasn't displaying the correct message.